Added final builds and adjusted PATH variable
Added newlib and the final gcc builds.

The PATH variable is now ENV instead of ARG.
PATH is fixed given the PREFIX. If the user wants to change the path of
the toolchain, he must change PREFIX, not PATH. For this, there is no
need for PATH to be ARG.
Also, the Linux built-in PATH environment variable overlaps the ARG PATH
variable defined in the Dockerfile, so that the path of the toolchain
was never included in the PATH environment variable, leading to errors
building newlib.